Move db-based meatadata store to cmd pkg
