+++ cmd/containerd-stargz-grpc/db/reader.go
