+++ fs/layer/layer_test.go
@@ -23,234 +23,15 @@
 package layer
 
 import (
-	"context"
-	"io"
-	"io/ioutil"
-	"net/http"
-	"path"
-	"path/filepath"
-	"strings"
 	"testing"
 	"time"
 
-	"github.com/containerd/containerd/reference"
-	"github.com/containerd/stargz-snapshotter/cache"
-	"github.com/containerd/stargz-snapshotter/estargz"
-	"github.com/containerd/stargz-snapshotter/fs/reader"
-	"github.com/containerd/stargz-snapshotter/fs/remote"
-	"github.com/containerd/stargz-snapshotter/fs/source"
-	"github.com/containerd/stargz-snapshotter/metadata"
-	"github.com/containerd/stargz-snapshotter/task"
-	"github.com/containerd/stargz-snapshotter/util/testutil"
-	digest "github.com/opencontainers/go-digest"
-	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
+	memorymetadata "github.com/containerd/stargz-snapshotter/metadata/memory"
 )
 
-const (
-	sampleChunkSize = 3
-	sampleData1     = "0123456789"
-	sampleData2     = "abcdefghij"
-)
-
-var testStateLayerDigest = digest.FromString("dummy")
-
-// Tests prefetch method of each stargz file.
-func TestPrefetch(t *testing.T) {
-	testPrefetch(t, "db", newDBReader)
-	testPrefetch(t, "mem", newMemoryReader)
+func TestLayer(t *testing.T) {
+	TestSuiteLayer(t, memorymetadata.NewReader)
 }
-
-func testPrefetch(t *testing.T, name string, factory readerFactory) {
-	defaultPrefetchSize := int64(10000)
-	landmarkPosition := func(t *testing.T, l *layer) int64 {
-		if l.r == nil {
-			t.Fatalf("layer hasn't been verified yet")
-		}
-		if id, _, err := l.r.Metadata().GetChild(l.r.Metadata().RootID(), estargz.PrefetchLandmark); err == nil {
-			offset, err := l.r.Metadata().GetOffset(id)
-			if err != nil {
-				t.Fatalf("failed to get offset of prefetch landmark")
-			}
-			return offset
-		}
-		return defaultPrefetchSize
-	}
-	tests := []struct {
-		name             string
-		in               []testutil.TarEntry
-		wantNum          int      // number of chunks wanted in the cache
-		wants            []string // filenames to compare
-		prefetchSize     func(*testing.T, *layer) int64
-		prioritizedFiles []string
-	}{
-		{
-			name: "no_prefetch",
-			in: []testutil.TarEntry{
-				testutil.File("foo.txt", sampleData1),
-			},
-			wantNum:          0,
-			prioritizedFiles: nil,
-		},
-		{
-			name: "prefetch",
-			in: []testutil.TarEntry{
-				testutil.File("foo.txt", sampleData1),
-				testutil.File("bar.txt", sampleData2),
-			},
-			wantNum:          chunkNum(sampleData1),
-			wants:            []string{"foo.txt"},
-			prefetchSize:     landmarkPosition,
-			prioritizedFiles: []string{"foo.txt"},
-		},
-		{
-			name: "with_dir",
-			in: []testutil.TarEntry{
-				testutil.Dir("foo/"),
-				testutil.File("foo/bar.txt", sampleData1),
-				testutil.Dir("buz/"),
-				testutil.File("buz/buzbuz.txt", sampleData2),
-			},
-			wantNum:          chunkNum(sampleData1),
-			wants:            []string{"foo/bar.txt"},
-			prefetchSize:     landmarkPosition,
-			prioritizedFiles: []string{"foo/", "foo/bar.txt"},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name+"-"+name, func(t *testing.T) {
-			sr, dgst, err := testutil.BuildEStargz(tt.in,
-				testutil.WithEStargzOptions(
-					estargz.WithChunkSize(sampleChunkSize),
-					estargz.WithPrioritizedFiles(tt.prioritizedFiles),
-				))
-			if err != nil {
-				t.Fatalf("failed to build eStargz: %v", err)
-			}
-			blob := newBlob(sr)
-			mcache := cache.NewMemoryCache()
-			mr, closeFn, err := factory(sr)
-			if err != nil {
-				t.Fatalf("failed to create metadata reader: %v", err)
-			}
-			defer mr.Close()
-			defer closeFn()
-			vr, err := reader.NewReader(mr, mcache, digest.FromString(""))
-			if err != nil {
-				t.Fatalf("failed to create reader: %v", err)
-			}
-			l := newLayer(
-				&Resolver{
-					prefetchTimeout:       time.Second,
-					backgroundTaskManager: task.NewBackgroundTaskManager(10, 5*time.Second),
-				},
-				ocispec.Descriptor{Digest: testStateLayerDigest},
-				&blobRef{blob, func() {}},
-				vr,
-			)
-			if err := l.Verify(dgst); err != nil {
-				t.Errorf("failed to verify reader: %v", err)
-				return
-			}
-			prefetchSize := int64(0)
-			if tt.prefetchSize != nil {
-				prefetchSize = tt.prefetchSize(t, l)
-			}
-			if err := l.Prefetch(defaultPrefetchSize); err != nil {
-				t.Errorf("failed to prefetch: %v", err)
-				return
-			}
-			if blob.calledPrefetchOffset != 0 {
-				t.Errorf("invalid prefetch offset %d; want %d",
-					blob.calledPrefetchOffset, 0)
-			}
-			if blob.calledPrefetchSize != prefetchSize {
-				t.Errorf("invalid prefetch size %d; want %d",
-					blob.calledPrefetchSize, prefetchSize)
-			}
-			if cLen := len(mcache.(*cache.MemoryCache).Membuf); tt.wantNum != cLen {
-				t.Errorf("number of chunks in the cache %d; want %d: %v", cLen, tt.wantNum, err)
-				return
-			}
-
-			lr := l.r
-			if lr == nil {
-				t.Fatalf("failed to get reader from layer: %v", err)
-			}
-			for _, file := range tt.wants {
-				id, err := lookup(lr.Metadata(), file)
-				if err != nil {
-					t.Fatalf("failed to lookup %q: %v", file, err)
-				}
-				e, err := lr.Metadata().GetAttr(id)
-				if err != nil {
-					t.Fatalf("failed to get attr of %q: %v", file, err)
-				}
-				wantFile, err := lr.OpenFile(id)
-				if err != nil {
-					t.Fatalf("failed to open file %q", file)
-				}
-				blob.readCalled = false
-				if _, err := io.Copy(ioutil.Discard, io.NewSectionReader(wantFile, 0, e.Size)); err != nil {
-					t.Fatalf("failed to read file %q", file)
-				}
-				if blob.readCalled {
-					t.Errorf("chunks of file %q aren't cached", file)
-					return
-				}
-			}
-		})
-	}
-}
-
-func lookup(r metadata.Reader, name string) (uint32, error) {
-	name = strings.TrimPrefix(path.Clean("/"+name), "/")
-	if name == "" {
-		return r.RootID(), nil
-	}
-	dir, base := filepath.Split(name)
-	pid, err := lookup(r, dir)
-	if err != nil {
-		return 0, err
-	}
-	id, _, err := r.GetChild(pid, base)
-	return id, err
-}
-
-func chunkNum(data string) int {
-	return (len(data)-1)/sampleChunkSize + 1
-}
-
-func newBlob(sr *io.SectionReader) *sampleBlob {
-	return &sampleBlob{
-		r: sr,
-	}
-}
-
-type sampleBlob struct {
-	r                    *io.SectionReader
-	readCalled           bool
-	calledPrefetchOffset int64
-	calledPrefetchSize   int64
-}
-
-func (sb *sampleBlob) Authn(tr http.RoundTripper) (http.RoundTripper, error) { return nil, nil }
-func (sb *sampleBlob) Check() error                                          { return nil }
-func (sb *sampleBlob) Size() int64                                           { return sb.r.Size() }
-func (sb *sampleBlob) FetchedSize() int64                                    { return 0 }
-func (sb *sampleBlob) ReadAt(p []byte, offset int64, opts ...remote.Option) (int, error) {
-	sb.readCalled = true
-	return sb.r.ReadAt(p, offset)
-}
-func (sb *sampleBlob) Cache(offset int64, size int64, option ...remote.Option) error {
-	sb.calledPrefetchOffset = offset
-	sb.calledPrefetchSize = size
-	return nil
-}
-func (sb *sampleBlob) Refresh(ctx context.Context, hosts source.RegistryHosts, refspec reference.Spec, desc ocispec.Descriptor) error {
-	return nil
-}
-func (sb *sampleBlob) Close() error { return nil }
 
 func TestWaiter(t *testing.T) {
 	var (
